+++ android/src/newarch/MetadataRetrieverSpec.kt
@@ -1,7 +1,0 @@
-package com.cyanchill.missingcore.metadataretriever
-
-import com.facebook.react.bridge.ReactApplicationContext
-
-abstract class MetadataRetrieverSpec internal constructor(context: ReactApplicationContext) :
-  NativeMetadataRetrieverSpec(context) {
-}