+++ example/src/data/useTracksWithSavedArtwork.ts
@@ -24,7 +24,9 @@ async function getTracksWithSavedArtwork() {
 
   const audioFiles = await getAudioFiles();
   console.log(
-    `Got list of audio files in ${((performance.now() - start) / 1000).toFixed(4)}s.`
+    `Got list of audio files in ${((performance.now() - start) / 1000).toFixed(
+      4
+    )}s.`
   );
 
   const assetURIMap = Object.fromEntries(
@@ -43,7 +45,10 @@ async function getTracksWithSavedArtwork() {
     })
   );
   console.log(
-    `Got metadata of ${audioFiles.length} tracks in ${((performance.now() - start) / 1000).toFixed(4)}s.`
+    `Got metadata of ${audioFiles.length} tracks in ${(
+      (performance.now() - start) /
+      1000
+    ).toFixed(4)}s.`
   );
   console.log('Errors:', results.errors);