+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ## [Unreleased]
 
+### ❗ Breaking Changes
+
+- Drop support for Old Architecture.
+
+### ⚙️ Internal Changes
+
+- Rewrite app to use Turbo Native Modules.
+
 ## [2.5.0] - 2026-02-04
 
 ### 🎉 Added